+++ zk/realtime/kafka/consumer.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/IBM/sarama"
 	"github.com/ledgerwatch/erigon/eth/ethconfig"
-	kafkaTypes "github.com/ledgerwatch/erigon/zk/kafka/types"
+	kafkaTypes "github.com/ledgerwatch/erigon/zk/realtime/kafka/types"
 	"github.com/ledgerwatch/log/v3"
 )
