+++ zk/realtime/kafka/types/error.go
